Move functions to their own lib
