+++ ca/certificate-authority.go
@@ -35,6 +35,7 @@ import (
 
 	"github.com/letsencrypt/boulder/cmd"
 	"github.com/letsencrypt/boulder/core"
+	csrlib "github.com/letsencrypt/boulder/csr"
 	blog "github.com/letsencrypt/boulder/log"
 )
 
@@ -381,8 +382,8 @@ func (ca *CertificateAuthorityImpl) IssueCertificate(ctx context.Context, csr x5
 
 	// Since the RPC layer uses the CertificateRequest.Raw field when transporting the CSR
 	// we must re-normalize the CN/DNSNames
-	core.NormalizeCSR(&csr, ca.forceCNFromSAN)
-	if err := core.VerifyCSR(&csr, ca.maxNames, &ca.keyPolicy, ca.PA, regID); err != nil {
+	csrlib.NormalizeCSR(&csr, ca.forceCNFromSAN)
+	if err := csrlib.VerifyCSR(&csr, ca.maxNames, &ca.keyPolicy, ca.PA, regID); err != nil {
 		// AUDIT[ Certificate Requests ] 11917fa4-10ef-4e0d-9105-bacbe7836a3c
 		ca.log.AuditErr(err)
 		return emptyCert, core.MalformedRequestError(err.Error())
